pyproject.toml
[tool.poetry]
name = "dangerzone"
version = "0.1.0"
description = "Take potentially dangerous PDFs, office documents, or images and convert them to safe PDFs"
authors = ["Micah Lee <micah.lee@theintercept.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
click = "*"
appdirs = "*"
PySide2 = "5.15.2.1"
colorama = "*"
pyxdg = {version = "*", platform = "linux"}
openai-whisper = "^20230124"

[tool.poetry.scripts]
dangerzone = 'dangerzone:main'
dangerzone-container = 'dangerzone:main'

# Dependencies required for packaging the code on various platforms.
[tool.poetry.group.package.dependencies]
setuptools = "*"
cx_freeze = {version = "^6.13.1", platform = "win32"}
pywin32 = {version = "*", platform = "win32"}
pyinstaller = {version = "*", platform = "darwin"}

# Dependencies required for linting the code.
[tool.poetry.group.lint.dependencies]
black = "*"
isort = "*"
mypy = "*"
PySide2-stubs = "*"

# Dependencies required for testing the code.
[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"
pytest-mock = "^3.10.0"
pytest-xdist = "^2.5.0"
pytest-cov = "^3.0.0"
strip-ansi = "*"

[build-system]
requires = ["poetry>=1.1.4"]
build-backend = "poetry.masonry.api"
